+++ src/components/ClientLeaderboard.astro
@@ -1,10 +1,7 @@
 
 <body>
     <div class="leaderboard">
-        <h1>Leaderboard</h1>
-        <div class="content">
-            <!-- Leaderboard Items Added Here -->
-        </div>
+        <h1>Please Sign In To View Leaderboard</h1>
     </div>
 </body>
