+++ src/components/Leaderboards.astro
@@ -19,4 +19,4 @@ if (Astro.cookies.has("__session")) {
 }
 ---
 
-{ loggedIn ? <LeaderboardServer/> : <LeaderboardClient /> }
+{ loggedIn ? <LeaderboardServer/> : <LeaderboardClient/> }
